Added score ranges to actions as well, making it possible to modify their scores directly

+++ src/main/kotlin/eater/ai/GenericActionWithState.kt
@@ -7,14 +7,14 @@ import eater.core.engine
 import kotlin.reflect.KClass
 
 class GenericActionWithState<T: Component>(name: String,
-                                           private val scoreFunction: (entity: Entity) -> Double,
+                                           private val scoreFunction: (entity: Entity) -> Float,
                                            private val abortFunction: (entity: Entity) -> Unit = {},
                                            private val actFunction: (entity: Entity, state: T, deltaTime:Float) -> Unit,
                                            private val componentClass: KClass<T>): AiAction(name) {
     lateinit var state: T
     val mapper = ComponentMapper.getFor(componentClass.java)
 
-    override fun score(entity: Entity): Double {
+    override fun score(entity: Entity): Float {
         return scoreFunction(entity)
     }
